requirements.txt
flask==3.1.0
apscheduler==3.10.4
google-api-python-client==2.160.0
google-auth==2.38.0
requests==2.32.3
phonenumbers==8.13.52
python-dotenv==1.0.1
gunicorn==23.0.0
pytz==2024.1
anthropic>=0.25.0
